+++ src/container/rx-module.ts
@@ -1,14 +1,25 @@
-import { Observable, Subject } from "rxjs";
-import { takeUntil } from "rxjs/operators";
-import { IModuleDestroy, IModuleHook, Module } from "./module";
+import { BehaviorSubject, Observable, Subject } from "rxjs";
+import { filter, takeUntil } from "rxjs/operators";
+import { IModuleDestroy, IModuleHook, IModuleOptions, Module } from "./module";
 
-/** Module class with RxJS unsubscribe subject embedded for usability. */
-export class RxModule extends Module {
+/** Module class with RxJS unsubscribe and state for usability. */
+export class RxModule<S = void> extends Module {
   /** Default module name. */
   public static readonly moduleName: string = "RxModule";
 
-  /** Internal observable management subject. */
-  private readonly rxUnsubscribe$ = new Subject<void>();
+  /** Observable state with takeUntil, filter pipe applied. */
+  public readonly rxState$: Observable<S>;
+
+  /** Internal observable subscription management. */
+  protected readonly rxUnsubscribe$ = new Subject<void>();
+
+  /** Internal observable state management. */
+  protected readonly rxStateRaw$ = new BehaviorSubject<S>(undefined as any);
+
+  public constructor(options: IModuleOptions) {
+    super(options);
+    this.rxState$ = this.rxTakeUntilModuleDown(this.rxStateRaw$).pipe(filter((x) => x != null));
+  }
 
   public moduleDown(...args: IModuleHook[]) {
     return super.moduleDown(...args, async () => {
@@ -19,6 +30,7 @@ export class RxModule extends Module {
   public moduleDestroy(...args: IModuleDestroy[]) {
     return super.moduleDestroy(...args, () => {
       this.rxUnsubscribe$.complete();
+      this.rxStateRaw$.complete();
     });
   }